Crate aptos_rest_client

Re-exports

pub use faucet::FaucetClient;

pub use response::Response;

pub use aptos_api_types;

pub use state::State;

pub use types::deserialize_from_prefixed_hex_string;

pub use types::Account;

pub use types::Resource;

Modules

aptos

error

faucet

response

state

types

Structs

Client

GasEstimationParams

IndexResponseBcs

The struct holding all data returned to the client by the index endpoint (i.e., GET “/”). This is just for the BCS response and cannot change

MoveModuleBytecode

Move module bytecode along with it’s ABI

PendingTransaction

A transaction waiting in mempool

VersionedNewBlockEvent

Enums

Transaction

Enum of the different types of transactions in Aptos

Constants

DEFAULT_VERSION_PATH_BASE

USER_AGENT

Functions

retriable

retriable_with_404